First pass at removing unused modules
First I checked by hand for modules that were not activated, not part of a bundle of modules of which some were activated, and with no obvious (or apparent) connections with modules that were activated. Then I found a drush command for checking algorithmically what was both inactive and not in the database, and could thus be removed by simply removing the folders (found at http://www.zyxware.com/articles/3887/drupal-how-to-remove-unused-modules-in-a-drupal-site-with-the-help-of-drush). I then removed only the modules that appeared on BOTH lists. Perhaps overly cautious, but for a first pass, brute-force mass removal, "overly cautious" is good.

Some of the hacks to contributed modules were to contributed moduels that we weren't using anymore, so this simplifies the update.