Reverting directory changes (more...)
I had seen something online that indicated that the moving the modules to the preferred subdirectory structure should work, with a couple of tweaks. It didn't work for me, and there are more important things for me to be spending my time on, so I'm just putting everything back directly under sites/all/modules.